Don't initialize variables with constants inside tf.while_loop
This was done in an attempt to work around "inputs from different
frames" error:

https://stackoverflow.com/questions/42564698/invalidargumenterror-the-node-has-inputs-from-different-frames

It turned out the problem is the serialization of tf.while_loop. When we
add the loss at training time (don't store it in the meta graph), the
error disappears. This might be a bug in tensorflow.

We keep this commit anyway.